+++ AutomaticRoleCheck.lua
@@ -1,33 +1,22 @@
-AutomaticRoleCheck = {}
-AutomaticRoleCheck.__index = AutomaticRoleCheck
-
-AutomaticRoleCheck.Defaults = {
-  Enabled = true,
-  DisableOnce = false,
-  DisableOnceOnLogin = false,
+AutomaticRoleCheck = {
+    FirstCheck = true,
+    Defaults = {
+      Enabled = true,
+      DisableOnce = false,
+      DisableOnceOnLogin = false,
+    }
 }
-AutomaticRoleCheck.Accept = function(self)
-  if AutomaticRoleCheck.Options.Enabled then
-    if not AutomaticRoleCheck.Options.DisableOnce then
-      self:Click()
-    end
-  end
-  AutomaticRoleCheck.Options.DisableOnce = false
+
+if not AutomaticRoleCheck_Options then
+  AutomaticRoleCheck_Options = AutomaticRoleCheck.Defaults
 end
-AutomaticRoleCheck.EventHandler = function(self, event, arg, ...)
-  if event == "ADDON_LOADED" and arg == "AutomaticRoleCheck" then
-    AutomaticRoleCheck.Options = AutomaticRoleCheck.Defaults
-    if AutomaticRoleCheck_Options then
-      for k, v in pairs(AutomaticRoleCheck_Options) do
-        AutomaticRoleCheck.Options[k] = v
-      end
-      AutomaticRoleCheck.Options.DisableOnce = AutomaticRoleCheck.Options.DisableOnce and AutomaticRoleCheck.Options.DisableOnce or AutomaticRoleCheck.Options.DisableOnceOnLogin
-    end
-    AutomaticRoleCheck.Panel.PopulatePanel()
-    AutomaticRoleCheck.Panel.Settings.PopulatePanel()
-  elseif event == "PLAYER_LOGOUT" then
-    AutomaticRoleCheck_Options = AutomaticRoleCheck.Options
+
+AutomaticRoleCheck.Accept = function(self)
+  if not (AutomaticRoleCheck_Options.DisableOnce or (AutomaticRoleCheck.FirstCheck and AutomaticRoleCheck_Options.DisableOnceOnLogin)) then
+    self:Click()
   end
+  AutomaticRoleCheck.FirstCheck = false
+  AutomaticRoleCheck_Options.DisableOnce = false
 end
 
 LFGListApplicationDialog.SignUpButton:HookScript("OnShow", AutomaticRoleCheck.Accept)